requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
Levenshtein==0.23.0
paramiko==3.3.1
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
PyMySQL==1.1.0
PyNaCl==1.5.0
python-dotenv==1.0.0
python-Levenshtein==0.23.0
PyYAML==6.0.1
rapidfuzz==3.5.2
sniffio==1.3.0
SQLAlchemy==2.0.23
sshtunnel==0.4.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
